+++ modules/Food_Service/Users/ActivityReport.php
@@ -47,10 +47,30 @@ if (UserStaffID() && ! $_REQUEST['modfunc'])
 			$RET[ $key ] = array($value) + $tmpRET;
 		}
 		//echo '<pre>'; var_dump($RET); echo '</pre>';
-		$columns = array('TRANSACTION_ID' => _('ID'),'FULL_NAME' => _('User'),'DATE' => _('Date'),'TIME' => _('Time'),'BALANCE' => _('Balance'),'DESCRIPTION' => _('Description'),'AMOUNT' => _('Amount'),'SELLER' => _('User'));
-		$group = array(array('TRANSACTION_ID'));
-		$link['remove']['link'] = PreparePHP_SELF($_REQUEST,array(),array('modfunc' => 'delete'));
-		$link['remove']['variables'] = array('transaction_id' => 'TRANS_ID','item_id' => 'ITEM_ID');
+
+		$columns = array(
+			'TRANSACTION_ID' => _( 'ID' ),
+			'FULL_NAME' => _( 'User' ),
+			'DATE' => _( 'Date' ),
+			'TIME' => _( 'Time' ), // TODO: use ProperDateTime().
+			'BALANCE' => _( 'Balance' ),
+			'DESCRIPTION' => _( 'Description' ),
+			'AMOUNT' => _( 'Amount' ),
+			'SELLER' => _( 'User' ),
+		);
+
+		$group = array( array( 'TRANSACTION_ID' ) );
+
+		$link['remove']['link'] = PreparePHP_SELF(
+			$_REQUEST,
+			array( 'delete_cancel' ),
+			array( 'modfunc' => 'delete' )
+		);
+
+		$link['remove']['variables'] = array(
+			'transaction_id' => 'TRANS_ID',
+			'item_id' => 'ITEM_ID',
+		);
 	}
 	else
 	{